refactor: extract FormatQuickBar to new file
